create the access button

+++ keyDown.php
@@ -4,9 +4,12 @@
  $keyCode = $_GET['keycode'];
  $filename = $_GET['filename'];
 
- $stringtext = file_get_contents('file/'.$filename);
- $stringtext = $stringtext.$keyCode;
- file_put_contents('file/'.$filename, $stringtext);
-
-	echo $stringtext;
+ if( is_file('file/'.$filename) ){
+ 	$stringtext = file_get_contents('file/'.$filename);
+ 	$stringtext = $stringtext.$keyCode;
+ 	file_put_contents('file/'.$filename, $stringtext);
+ 	echo $stringtext;
+ }else{
+ 	echo 'file don\'t exists!';
+ }
 ?>
